+++ lib/src/audit.dart
@@ -5,6 +5,9 @@ import 'dart:async';
 
 /// Creates a StreamTransformer which only emits once per [duration], at the
 /// end of the period.
+///
+/// Like `throttle`, except it always emits the most recently received event in
+/// a period.  Alwyas introduces a delay of at most [duration].
 StreamTransformer<T, T> audit<T>(Duration duration) {
   Timer timer;
   bool shouldClose = false;